Use double for Distance precision

+++ src/main/java/org/openstreetmap/atlas/geography/atlas/packed/PackedAtlas.java
@@ -1181,7 +1181,8 @@ public final class PackedAtlas extends AbstractAtlas
         {
             final Distance distance1 = location.distanceTo(node1.getLocation());
             final Distance distance2 = location.distanceTo(node2.getLocation());
-            final long difference = distance2.asMillimeters() - distance1.asMillimeters();
+            final long difference = (long) distance2.asMillimeters()
+                    - (long) distance1.asMillimeters();
             if (difference > 0)
             {
                 return 1;
